Made score dynamic
- The amount of points the player earns from destroying enemies now
depends on how far up the screen the enemy was killed. The closer to
the bottom the enemy was, the more points.
- Fixed incorrect wave spawn times
- There is now 10 seconds of inactivity before enemies start spawning.
A tutorial sequence should be added to fill up this time

What I’ll work on next:
- Implement Meghan’s ship model for the player
- Make the game show how many points the player earned when they
destroy an enemy
- Add more movement features (ex. SHIFT to speed up/slow down the
player’s movement)
- Either remove bullets colliding with each other (if it turns out to
be more annoying than interesting) or add some visual indicator showing
the collision of the bullets
- Add visual feedback for when the player gets damaged (can be as
simple as the model flashing red)
- Add background music
- Add sound effects